[cgl] Add glitz-cgl backend
The backend is mostly derived from the AGL backend, but allows creating
drawable on Cocoa NSView objects (*not* NSOpenGLView: the latter is
mostly a NSView + OpenGL contexts and pixel formats, which glitz manages
on its own).  All the added files are Objective-C (.m).

The only changes with respect to the AGL backend were in the pixel
format queries that: 1) need to use CGL directly in order to avoid
logging messages for unavailable formats; 2) needed some tweaking in
order to find no-depth, no-stencil pbuffer formats (the new pbuffer
format discovery code actually does not need it, but it's better
to be safe; 3) have a slightly different sort order that is nicer (as
seen in the output of glitzinfo).

The renderer test's Objective-C bits in glitz_cgl.m has two parts in it.
One is the boilerplate that sets up Cocoa without using .nib files;
it can be useful only if you want to write a similar example program
for Glitz.  I split the test-specific code in two (before parsing command
line/after parsing command line) to avoid showing the icon in the dock
if the program is going to exit very soon.

The second part is GlitzOpenGLView, which provides the bits of NSOpenGLView
that go beyond storing OpenGL contexts and pixel formats; see the "Custom
Cocoa OpenGL View" example from Apple.  Programs using Glitz in their UI
should be able to recycle it pretty easily as it's just four methods (one
caveat: I didn't investigate how to deal with window resizes).