[agl] Rewrite extension discovery code for AGL
This patch rewrites the code to discover extensions in the AGL backend.
In particular, pbuffer formats are discovered by creating a context and
selecting a pbuffer into it.  This does not make a difference for AGL,
but it does for CGL and I implemented it also here so that the code for
the two backends does not diverge too much.