[glx] Add glitz_glx_set_render_type method.
Add glitz_glx_set_render_type method to specifies whether rendering is to be
done with a direct connection to the graphics system if possible or through
the X server.
It can be set with LIBGL_ALWAYS_INDIRECT, but this method provide
multi-threading programs to have both context in same time.