[cairogears] add cube demo
This is the dual of the other opengl demo, as it uses Glitz just in the
backend to prepare the texture, and then uses normal OpenGL to draw
user graphics.

It does not work on the AGL backend because of the same "feature"
that causes different overlay surfaces to be created when multiple
AGL contexts refer to the same drawable.  To fix it, we could set
AGL_SURFACE_ORDER to -1 for the window drawables, *and* punch a
hole in the Carbon drawing so that the OpenGL graphics show through.

This would give, from bottom to top:
- the window drawable's overlay surface (AGL_SURFACE_ORDER == -1)
- the Carbon drawing
- the user contexts for framebuffer (but with an empty AGL_BUFFER_RECT)
- the user contexts for visible 3D stuff (on top of everything)